abort関数をfollowメソッド側に引っ越しさせ、条件に合わなかったときの処理を記述。 followメソッド内に既述したabort関数が条件に一致してDBの処理が終わった後に実行されないようにreturn;を追記。 followCheckメソッドのロジックをログインしているユーザーを基準としてリレーションメソッドを使う仕様に変更。 followメソッド内の条件分岐に利用した変数は不要だったので削除。 unfollowメソッドには当初条件分岐をつけないつもりだったが、followメソッドと全く同じ条件を追記し、return;とabort関数も併せて記述。 unfollowメソッドに条件分岐を加えたことで、ユーザー退会処理の際に条件に一致しないためunfollowメソッドを呼び出しできなくなったので、別途ユーザー退会のメソッドの中に、 退会時にフォローデータが削除されるロジックをリレーションメソッドを使って実装。 マイグレーションファイルの記述を直接カラムの生成に関係ない記述を一行を空けてまとめて記述した。 ユーザーを更新するルーティングを ルートパラメーターでユーザーidを受取れるURLに修正。 ルートパラメーターを送るためにedit.blade.phpファイルのroute関数の第二引数にユーザーidを指定。そして不要になったinputタグを削除。 UserController内のupdateUserメソッドでrouteから渡されるユーザーidを変数で受け取れるように二つ目の引数の変数を実装。

+++ app/User.php
@@ -75,16 +75,17 @@ class User extends Authenticatable
      */
     public function follow($id) {
         DB::BeginTransaction();
-        try {
-            $bool = false;
-            if (Auth::id() !== $id && !$bool = $this->followCheck($id)) {
+        if (Auth::id() !== $id && !$this->followCheck($id)) {
+            try {
                 $this->followings()->attach($id);
                 DB::commit();
+            } catch (\Throwable $e) {
+                DB::rollBack();
+                abort(500);
             }
-        } catch (\Throwable $e) {
-            DB::rollBack();
-            abort(500);
+            return;
         }
+        abort(404);
     }
 
     /**
@@ -93,11 +94,15 @@ class User extends Authenticatable
      * @return view
      */
     public function unfollow($id) {
-        try {
-            $this->followings()->detach($id);
-        } catch (\Throwable $e) {
-            abort(500);
+        if (Auth::id() !== $id && !$this->followCheck($id)) {
+            try {
+                $this->followings()->detach($id);
+            } catch (\Throwable $e) {
+                abort(500);
+            }
+            return;
         }
+        abort(404);
     }
 
     /**
@@ -106,7 +111,7 @@ class User extends Authenticatable
      * @return view
      */
     public function followCheck ($id) {
-        $user = User::find($id);
-        return $user->followers()->where('user_id', Auth::id())->exists();
+        $user = Auth::user();
+        return $user->followings()->where('followed_id', $id)->exists();
     }
 }
